+++ picomint-client/src/ln/mod.rs
@@ -45,10 +45,10 @@ use self::send_sm::{SendSMCommon, SendSMState, SendStateMachine, SendStateMachin
 /// Maximum total contract lock, in blocks, the client is willing to accept
 /// from a gateway. Backstop against an abusive gateway tying funds up before
 /// the unilateral refund path opens.
-const EXPIRY_DELTA_LIMIT: u64 = 1000;
+const EXPIRY_DELTA_LIMIT: u16 = 1000;
 
 /// A two hour buffer in case either the client or gateway go offline
-const CONTRACT_CONFIRMATION_BUFFER: u64 = 12;
+const CONTRACT_CONFIRMATION_BUFFER: u32 = 12;
 
 /// Contracts pulled per round trip when walking the incoming-contract
 /// stream.
@@ -224,7 +224,9 @@ async fn send_inner(
         payment_hash: *invoice.payment_hash(),
         amount,
         fee,
-        expiry: consensus_block_count + gateway_info.expiry_delta + CONTRACT_CONFIRMATION_BUFFER,
+        expiry: consensus_block_count
+            + u32::from(gateway_info.expiry_delta)
+            + CONTRACT_CONFIRMATION_BUFFER,
         claim_pk: gateway_info.module_public_key,
         refund_pk: refund_keypair.x_only_public_key().0,
     };
